+++ Demo/src/main/java/com/entornos/project/Demo/Service/impl/MedicamentoService.java
@@ -3,6 +3,7 @@ package com.entornos.project.Demo.Service.impl;
 import com.entornos.project.Demo.Model.Medicamento;
 import com.entornos.project.Demo.DTO.*;
 import com.entornos.project.Demo.Repository.MedicamentoRepository;
+import com.entornos.project.Demo.Service.interfaces.IMedicamentoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -10,7 +11,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class MedicamentoService {
+public class MedicamentoService implements IMedicamentoService {
 
     private final MedicamentoRepository repository;
 
@@ -21,44 +22,46 @@ public class MedicamentoService {
                 .collect(Collectors.toList());
     }
 
-    public MedicamentoDTO agregar(CrearMedicamentoDTO request) {
-        Medicamento medicamento = Medicamento.builder()
-                .nombre(request.getNombre())
-                .precio(request.getPrecio())
-                .imagenMed(request.getImagenMed())
-                .tipo(request.getTipo())
-                .build();
+    @Override
+    public MedicamentoDTO crear(CrearMedicamentoDTO crearMedicamentoDTO) {
+        Medicamento medicamento = new Medicamento(crearMedicamentoDTO);
         return toDTO(repository.save(medicamento));
     }
 
-    public MedicamentoDTO actualizar(Long id, CrearMedicamentoDTO request) {
-        Medicamento medicamento = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Medicamento no encontrado"));
+    @Override
+    public MedicamentoDTO actualizar(ActualizarMedicamentoDTO actualizarMedicamentoDTO) {
+        Medicamento medicamento = repository.findById(actualizarMedicamentoDTO.getId()).orElseThrow(() -> new RuntimeException("Medicamento no encontrado"));
 
-        medicamento.setNombre(request.getNombre());
-        medicamento.setPrecio(request.getPrecio());
-        medicamento.setImagenMed(request.getImagenMed());
-        medicamento.setTipo(request.getTipo());
+        medicamento.setNombre(actualizarMedicamentoDTO.getNombre());
+        medicamento.setPrecio(actualizarMedicamentoDTO.getPrecio());
+        medicamento.setImagen(actualizarMedicamentoDTO.getImagenMed());
+        medicamento.setVentaLibre(actualizarMedicamentoDTO.getVentaLibre());
 
         return toDTO(repository.save(medicamento));
     }
 
-    public void eliminar(Long id) {
+    @Override
+    public MedicamentoDTO eliminar(Long id) {
+        Medicamento medicamento = repository.findById(id).orElse(null);
+        if (medicamento == null) throw new RuntimeException("Medicamento no encontrado");
         repository.deleteById(id);
+        return toDTO(medicamento);
     }
 
     private MedicamentoDTO toDTO(Medicamento m) {
         return MedicamentoDTO.builder()
-                .id(m.getMedicamentoId())
+                .id(m.getId())
                 .nombre(m.getNombre())
                 .precio(m.getPrecio())
-                .tipo(m.getTipo())
+                .ventaLibre(m.getVentaLibre())
                 .build();
     }
 
+    @Override
     public MedicamentoDTO obtenerPorId(Long id) {
         Medicamento medicamento = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Medicamento no encontrado con ID: " + id));
+                .orElse(null);
+        if (medicamento == null) return null;
         return toDTO(medicamento);
     }
